+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+### Added
+- Added `State` property to `KafkaEventReceiver`
+
 ## [2.2.2] 2023-02-07
 ### Bugfix
 - Correctly cascade deletes from stream to schema when `entityView.exist.check.enabled` is `false`.